feat(mobile): add platform-aware OCR abstraction layer (S-43)
Refactor OCR service into a provider-based architecture for platform
abstraction. ML Kit provider handles native iOS/Android, web provider
stubs tesseract.js for Phase 6. Both conform to the OcrProvider
interface ensuring consistent output format across platforms.

- New: OcrProvider interface with recognizePage contract
- New: mlkitProvider — ML Kit implementation (extracted from S-42)
- New: webProvider — tesseract.js stub (Phase 6)
- New: types.ts — shared OCR type definitions
- Refactored: ocrService.ts dispatches to platform provider
- Public API unchanged (performOcr, performOcrBatch, etc.)

Closes #44
